src/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coverage==7.0.5
crypto==1.4.1
cryptography==39.0.0
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
mccabe==0.7.0
Naked==0.1.32
passlib==1.7.4
psycopg2-binary==2.9.5
pyAesCrypt==6.0.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
shellescape==3.8.1
sqlparse==0.4.3
urllib3==1.26.14